ProQuest error checking and re-submission
We have been seeing a bug where some ETDs submitted to ProQuest
don't have their binary attachments. AFAICT, re-running an export
of the same ETD results in a valid export, so there must be some kind
of intermittent problem (lag from fedora?).

During the building of the export zip file, check to ensure
the binary files exist and raise an exception if they don't.
This will cause the sidekiq job to fail and re-run until it works.

Also, add a rake task to export a specific object, and build in
parameters for file transmission to proquest, file cleanup, and
re-submission, so we can re-queue the objects that failed and
more easily troubleshoot future issues.