Don't generate derivatives for office document mime types
Office document mime types lead us to convert (large!) documents to .pdf before
generating a .jpg. This is resource intensive, and not necessary in this
application (we are always more concerned about the primary PDF content, and
could generate the .doc thumbnail from that if need).

We disable derivatives generation by making the relevant method a no-op.