Handle case where members exist, but not primary files
If a user creates an item with supplementary files but no primary files, the
view would break on `NoMethodError` for a `nil`. This is because we were
guarding on whether *any* members existed, but actually seeking a `#primary?`
FileSet member. In leiu of a larger refactor, we add a second guard to prevent
the primary file from displaying.

This case would also be triggered between processing of multiple
`AttachFileToWorkJob` jobs shortly after upload. Since the jobs are
asynchronous, it's possible for a supplementary file to be attached before the
primary--even when a primary file will eventually exist for the object.

Closes #921.