Use a factory method for department and subfield services
Hiding this behavior in the `EtdHelper` makes it hard to test. Using a factory
method allows us to easily check that a given `school` or `department` yields a
correct service result.

For now, the factory just uses a case statement. We could switch to a registry
or some other pattern for matching schools or departments to their respective
services.

Several services had been renamed, but not corrected in bcb83332. These are
fixed here and regression tests are added.

Connected to #924.