Setup `ImportUrlJob` (Box retrieval) to use the :ingest queue
Box retrieval is a high value task that runs the risk of timeout if
delayed. It's also not a resource intensive task on our side; network latency
probably accounts for most of the elapsed clock time. For this reason, it's a
good idea to put this job in a high priority queue.

We move it to `:ingest`. The other jobs there share similar properties. For
clarity, we set this in the `queues.rb` initializer, which is where we are
setting other queues originally set by Hyrax. Because we have a custom
`ImportUrlJob` that specifies its queue, we set it there too, to avoid
confusion. The initializer version runs later, and is canonical.

See #961.