Setup an 'Unassigned' `AdminSet`
The unassigned `AdminSet` is introduced as a catch-all for submitted ETDs that
couldn't be automatically assigned to another AdminSet. It has a generic
one-step approval process.

The purpose of this change is to avoid raising user facing errors when complex
internal admin set logic fails. Ideally, no items would ever be assigned to this
set; when they are, however, they can be triaged and repaired by
administrators. This avoids letting issues impact students at submission time.

Related to #924.