[Hyrax-2.0] Filter groups from approvers and reviewers for notification
Hyrax 2.0 adds the admin group to approvers and reviewers by default. This
uncovered a bug where we would fail to find a group included in a workflow
permission. This patches the bug by filtering based on proxy type for Sipity
agents. An accompanying refactor reduces the number of database interactions
required when setting up a notification.