Story 1029: Display config for schools
The configured relationships between schools, departments, subfields,
admin sets, approving users, and workflows is complicated.  That
information is spread across a dozen yaml files and 5 or 10 database
tables.  It's really hard to understand the current configuration for a
running app.

This code attempts to make it easier to understand the config by
displaying the full config on a single web page for each school.

Some things to note:

* I made a SchoolsController to display the config info for each school,
which is only accessible by admin users.

* I pulled the logic for choosing the correct AdminSet into a new class
AdminSetChooser, so that I could share that logic between the Etd class
and the School and Department classes.

* I copied app/views/hyrax/dashboard/sidebar/_configuration.html.erb over
from hyrax so that I could insert the nav link for the schools config
page.