Add webpacker and Vue
This commit adds webpacker and Vue to the codebase. Webpacker
is a way of managing front-end assets. Vue is a progressive
JS framework that is designed to be incrementally adoptable.

Additional applications need to be installed in the dev env
after this commit. You will need to install `yarn`, a JS package
manager similar to `gem`.

On a Mac you can install it like this:

`brew install yarn`

On other systems you can install it though `npm`:

`npm install -g yarn`

When running the dev env you will need to have the
`webpack-dev-server` running in another tab:

`bin/webpack-dev-server`

Co-authored-by: Jennifer Linder <jenlindner@gmail.com>
Co-authored-by: Jamie Little <jamie@jamielittle.org>