Fixed error when click 'All Collections' link. Story #1204
There are 2 parts to this fix:

1. The hyrax view assumes that the description returned by the presenter
is an array (and attempts to call `.first` on it).  That will cause an
error in the case where description is nil.  So I made a change to the
SolrDocument#description to return empty array if description is nil to
avoid raising an error.

2. In the case where the Collection does have a description, the
presenter/SolrDocument was returning nil.  This is because the
Collection model was using an indexer that doesn't index the
BasicMetadata.  (We missed a new indexer during our recent hyrax gem
upgrades.)