Add testing and linting for scripts in app/javascript
This commit sets up tesing and linting for the code in
app/javascript. The tests can be run with  `yarn test` and
linting can be run with `yarn lint`.

Testing is done with http://jestjs.io/ and can be used with
modern JS modules, Vue Single File Components, or plain JS.
There is an inital
test in app/javascript/test/app.spec.js that can be used
as a reference. The assertions that be made are documented here:
http://jestjs.io/docs/en/expect

The travis build will run `yarn test`.

Linting is done with the standard js style: https://standardjs.com/

Travis is not running `yarn lint` yet. There will need to be some
clean up of the exisiting code before making that part of the build
process if we want to make it part of the build process.

Related to #1181