Add dynamic Submission Type dropdown
This commit adds a compontent that renders the
the submission types dropdown on the form. It
does an XHR request to the QA endpoint to accomplish
this.

There is a basic spec in place, but to fully test
the component, there will need to be an Axios mock
in place. The spec currently triggers a node warning.