Add dynamic subfield based on the selected department
This changes the `formStore` so that it has the mappings
required to determine what vocab should be looked up
to determine the subfields for a chosen department.

If additional subfields need to be included, they should
be placed in the `formStore`.

There is a subfield component that renders the in the
'My Program' tab.