Basic uploading of supplementary files
This commit adds basic supplementary file
functionality. It doesn't include the metadata
inputs needed for the form. Those will need to be
added later.

Right now the delete action for the supplementary files
isn't working correctly (I assume that there is a different
controller that deletes those files?)

This also changes some of the state management for drop-downs
in the form. Select boxes need to be handled a little bit
differently than the other input components.

q