Update and test `CreateWithRemoteFilesActor` (#1342)
* Update and test `CreateWithRemoteFilesActor`

This actor is modified from the Hyrax version. It has previously been untested,
and was significantly out of date. The status quo would have failed anytime a
remote file was sent to the Actor Stack. This updates the actor to reflect
changes in the Hyrax version and puts in rudimenary testing.

* Test setting `#pcdm_use` on remote files

Adds tests for custom `pcdm_use` behavior in `CreateWihRemoteFilesActor`.

* Optimize pcdm_use setting

Use `#find_by` instead of `where`. Also, don't save twice in a row.