Fix upload of supplemental files from box
The ability to attach metadata to browse-everything
uploaded supplemental files seems to have broken
during a recent upgrade, which included a change requiring
params to be whitelisted. This PR allows us to fetch
the required params hash even if they have not been specifically
whitelisted. Ideally, we would whitelist them, but afaict
the names of the params are not predictable.

Unfortunately, we don't have automated testing for browse
everything because it requires integration with an
external service.