Workaround ActiveFedora's non-persistence of nested attributes
ActiveFedora fails to persist when using indexed nested attributes like:

    { my_property_attributes: { '0' => { name: ['moomin'] } } }

The library does not propagate changes made in this way to `#resource` which
prevents them from being added to the `ActiveFedora::ChangeSet` and sent to
Fedora. As a work-around, we call `persist!` manually on known nested attributes
in the actor stack. This ensures that updates made through the form (and only
those made through the form) are properly saved.

We also change an existing test to document a workable procedure for changing
nested attributes directly using the models.

Fixes #1376