Add `VisibilityTranslator` as alternate collaborator for `#visibiliy`
We want to extend `Etd#visibility` to provide new visibility levels:

   - `VisibilityTranslator::FILES_EMBARGOED`: indicates that the work is
     under embargo, but that it is discoverable (not `restricted`/private).
     FileSets that are members of the work should be under a parallel embargo,
     making tthem private until the embargo end date.
   - `VisibilityTranslator::TOC_EMBARGOED`: indicates that the work is
     under embargo, as described above, and that the table of contents should
     be suppressed for the duration of the embargo.
   - `VisibilityTranslator::ALL_EMBARGOED`: indicates that the work is
     under embargo, as described above, and that the table of contents *and*
     abstract should be suppressed for the duration of the embargo.

This adds a `VisibilityTranslator` interface and base implementation (as
described above) that can be used as a replacement for the default `#visibility`
method on any `Hydra::AccessControl::Visibility` object. Objects implementing
this interface are intended to be injected as collaborators to allow swappable
`#visibility` implementations.

The end goal of this work is to allow administrators and approvers to edit
existing embargoes. This requires being able to select an appropriate visibility
as the "during embargo" and "after embargo" values at edit time.

Connected to #1206.