Allow approvers to read and edit Etds and FileSets
This ensures approvers can read (download) and edit Works and FileSets. Since
there's not an explicit "approving" group that we can use to set ACLs for these
works, we need to set the `can` permissions explictly. Whenever
`can_review_submissions?` is `true`, we let users read and edit a work, as well
as forcing `test_download` to `true`.

Connected to #1206, #1216.