Set `FileSet` visibility and embargo on work embargo edit
Our file set visibility behavior is somewhat different from Hyrax's
default behavior. When we update work visibility or permissions, we
want to ensure that the FileSets are given permissions congruent with
the embargo settings. We do this by setting the FileSet visibility using
 `FileSetActor` directly, inline.

We override the `EtdsController#after_update_response` method to avoid
redirects to permission edit confirmation pages.

Embargoes and visibility can be set inline, since we expect the number of
`FileSet`s per work to be limited to a small handful.

Closes #1209.