When displaying active embargoes search only Etd embargoes
The purpose of this filter is to avoid showing FileSet/File embargoes when
showing the embargo index.

The normal reason to view the embargo index is to edit embargoes from the
dashboard. In the usual case for this application we don't want users editing
embargoes for FileSets/Files, since these are meant to be derived from the
embargo on the `Etd`. Indeed, we can safely consider file embargoes to be
"behind the scenes"; they have more to do with permission management than user
facing funcionality. To avoid confusion, we simply decline to list these
embargoes in the index view.

If additional, non-`Etd` work types are added, this may need to be updated to
filter in a different way.

Closes #1404.