Part of story #1445 - allow student to replace their primary PDF.
This code handles the case where the student has already submitted an
ETD, and they use the 'Remove' button to delete their primary PDF file.

I added code to the refresh_from_etd method to include the data that the
JavaScript needs to display the PDF file on the form along with the
'Remove' button to delete that file.

Note that since the student has already submitted their ETD, an 'Etd'
record with attached 'FileSet' record(s) has already been created, so
when the student hits the 'Remove' button, the delete action will go to
the FileSetsController instead of the UploadsController (which is where
it would go if the FileSet(s) hadn't been created yet).