Part of story #1445 - allow student to replace primary file (#1581)
This code fixes a problem when the student had an existing primary PDF
file, and then tried to add a supplemental file, there would be a 404
error.

The problem was caused by the fact that the 'uploaded_files' input on
the form contained not just the newly-uploaded file record IDs, but also
the ID of the previously saved primary PDF.  So when the controller
received that extra ID, it would look for the corresponding
Hyrax::UploadedFiles record (which doesn't exist).

I added a conditional to the hidden 'uploaded_files' input that prevents
previously-saved files from being added to the array.  Previously-saved
files can be differentiated from newly-uploaded files by the deleteUrl,
which contains the model name.