Display embargo type in edit screen
This introduces a class that translates the embargo
attributes back into a string that be editied on the form.

When you save from an edit, I would assume that the
embargo code in the Etd actor would translate the
string back into the attributes, but this isn't happening.

Related to #1595