Better management of school, dept, and embargo state
This commit fixes some bugs in the management of the
state of the properties that are based on the school.

The state is based on the saved school primarily now.

A copy of the school is also saved in the browser's
`localStorage` as a precaution.

This commit also adds a  destroy action to to the
InProgressEtd controller, this allows for a Start
Over button to be in the form that can be used
by a user to delete their InProgressEtd and start
over from scratch in case something goes wrong or
they enter the wrong school.

For #1623, #1622, #1603