Persist primary file across page refresh
The primary files is special case because
it needs to be persisted even if a user
navigates to Box even if they have not
saved the tab.

Box files pesist when tabbing around, but not if you navigate away from
the page without saving. That is how
all of the other tabs function! Primary
files are special case.

This persists the primary file JSON in localStorage so that if
the user moves to Box or to
another tab, the primary file remains.

This also keeps the Box UI present
after a file has been picked so that
multiple files can be added.

Connected to #1627
Connected to #1618
Contented to #1617