Display the supplemental files on the ETD edit form.  Part of story #1445.
When editing an ETD that I have previously submitted, the supplemental
files that I previously attached to my ETD will be displayed on the edit
form, along with the button to add more supplemental files.

There is a 'Remove' button next to each supplemental file in the list,
so I can delete any of the supplemental files.

For supplemental files that were already attached to the ETD, I won't be
able to edit the metadata fields (inputs will be disabled).  But if I
upload a new supplemental file, I can enter metadata for that new file.
(Note: Currently that new metadata won't be persisted when you hit the
submit button.  This is a known problem.  I will follow up later with a
separate PR for that problem.)