Fixed a problem where submitting the form for a previously-submitted Etd record would result in an infinite loop of patch updates to the EtdsController.
Because there are several forms in the UI that will hit
EtdsController#update, we need to differentiate between requests that
come from the edit form for the ETD and some other form (such as embargo
edit form).  I added a hidden field 'request_from_form' to make it
possible to clearly distinguish which form sent the request.

If the request came from the ETD edit form, then return a json response
that tells the form to redirect to the show page for the Etd record
after a successful update.