Do not allow unauthorized users to load edit form
Previously, unauthenticated users could load an edit
form, but they couldn't submit it.
This adds an extra layer of security to ensure that
unauthenticated users who attempt to load an edit
form for an in progress ETD are redirected to the
main page.
Authenticated users who attempt to load an edit form
that isn't theirs were already getting an "Unauthorized"
message, and this behavior has been left the same.
Authenticated users who attempt to load an edit
form for a nonexistent in progress ETD are also
redirected to the main page to prevent exceptions
being thrown.